requirements.txt
Flask==3.1.3
Flask-WTF==1.2.1
bcrypt==4.1.2
Flask-MySQLdb==2.0.0
mysqlclient==2.2.4
twilio==9.0.5
joblib==1.4.2
scikit-learn==1.5.2
pandas==2.2.2
numpy==1.26.4